Add helpers that generate scripts for copying files to unpacked gpu traces
The source debugging in a GPU trace often doesn't because of include paths (with extra complexity from strip_include_prefix, include_prefix).

This change adds the ability to generate scripts that take the temp dir that Xcode unpacks a gputrace into and where it recompiles the shaders. It adds symlinks to the included files in the bazel workspace so that the compile can succeed and source debugging of shaders works.